+++ README.md
@@ -1,2 +1,6 @@
 # vcc_juniorAPI
+![Python](https://img.shields.io/badge/python-3.8-blue.svg) [![API tests](https://github.com/SanjinDedic/vcc_senior/actions/workflows/Tests.yml/badge.svg)](https://github.com/SanjinDedic/vcc_senior/actions/workflows/Tests.yml)
+
+
+
 API for the Junior Years of the Victorian Coding challenge
